backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
jinja2==3.1.4
python-dotenv==1.0.1
aiofiles==24.1.0
pillow==10.4.0
python-barcode==0.16.1
pyserial==3.5
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
